+++ wemake_python_styleguide/visitors/tokenize/primitives.py
@@ -27,10 +27,6 @@ def _replace_braces(string: str) -> str:
 class WrongNumberTokenVisitor(BaseTokenVisitor):
     """Visits number tokens to find incorrect usages."""
 
-    _bad_number_suffixes: ClassVar[Pattern[str]] = re.compile(
-        r'^[0-9\.]+[BOXE]',
-    )
-
     _leading_zero_pattern: ClassVar[Pattern[str]] = re.compile(
         r'^[0-9\.]+([box]|e\+?\-?)0.+', re.IGNORECASE | re.ASCII,
     )
@@ -41,12 +37,6 @@ class WrongNumberTokenVisitor(BaseTokenVisitor):
     _positive_exponent_patterns: ClassVar[Pattern[str]] = re.compile(
         r'^[0-9\.]+e\+', re.IGNORECASE | re.ASCII,
     )
-
-    _bad_hex_numbers: ClassVar[FrozenSet[str]] = frozenset((
-        'a', 'b', 'c', 'd', 'e', 'f',
-    ))
-
-    _bad_complex_suffix: ClassVar[str] = 'J'
 
     _float_zero: ClassVar[Pattern[str]] = re.compile(
         r'^0\.0$',
@@ -59,20 +49,10 @@ class WrongNumberTokenVisitor(BaseTokenVisitor):
         Regressions:
         https://github.com/wemake-services/wemake-python-styleguide/issues/557
         """
-        self._check_complex_suffix(token)
         self._check_underscored_number(token)
         self._check_partial_float(token)
         self._check_bad_number_suffixes(token)
         self._check_float_zeros(token)
-
-    def _check_complex_suffix(self, token: tokenize.TokenInfo) -> None:
-        if self._bad_complex_suffix in token.string:
-            self.add_violation(
-                consistency.BadComplexNumberSuffixViolation(
-                    token,
-                    text=self._bad_complex_suffix,
-                ),
-            )
 
     def _check_underscored_number(self, token: tokenize.TokenInfo) -> None:
         if '_' in token.string:
@@ -90,11 +70,6 @@ class WrongNumberTokenVisitor(BaseTokenVisitor):
             )
 
     def _check_bad_number_suffixes(self, token: tokenize.TokenInfo) -> None:
-        if self._bad_number_suffixes.match(token.string):
-            self.add_violation(
-                consistency.BadNumberSuffixViolation(token, text=token.string),
-            )
-
         float_zeros = self._leading_zero_float_pattern.match(token.string)
         other_zeros = self._leading_zero_pattern.match(token.string)
         if float_zeros or other_zeros:
@@ -112,19 +87,6 @@ class WrongNumberTokenVisitor(BaseTokenVisitor):
                     text=token.string,
                 ),
             )
-
-        if token.string.startswith('0x') or token.string.startswith('0X'):
-            has_wrong_hex_numbers = any(
-                char in self._bad_hex_numbers
-                for char in token.string
-            )
-            if has_wrong_hex_numbers:
-                self.add_violation(
-                    consistency.WrongHexNumberCaseViolation(
-                        token,
-                        text=token.string,
-                    ),
-                )
 
     def _check_float_zeros(self, token: tokenize.TokenInfo) -> None:
         if self._float_zero.match(token.string):
@@ -169,11 +131,6 @@ class _StringTokenChecker:
         token: tokenize.TokenInfo,
         modifiers: str,
     ) -> None:
-        if 'u' in modifiers.lower():
-            self._add_violation(
-                consistency.UnicodeStringViolation(token, text=token.string),
-            )
-
         for modifier in modifiers:
             if modifier in self._bad_string_modifiers:
                 self._add_violation(