+++ wemake_python_styleguide/visitors/tokenize/statements.py
@@ -1,33 +1,12 @@
 import tokenize
 from collections import defaultdict
 from operator import attrgetter
-from typing import (
-    ClassVar,
-    DefaultDict,
-    Dict,
-    List,
-    Mapping,
-    Optional,
-    Sequence,
-    Tuple,
-)
+from typing import DefaultDict, List, Optional
 
 from typing_extensions import TypeAlias, final
 
-from wemake_python_styleguide.logic.tokens.brackets import (
-    get_reverse_bracket,
-    last_bracket,
-)
 from wemake_python_styleguide.logic.tokens.comprehensions import Compehension
-from wemake_python_styleguide.logic.tokens.constants import (
-    ALLOWED_EMPTY_LINE_TOKENS,
-)
-from wemake_python_styleguide.logic.tokens.constants import (
-    MATCHING_BRACKETS as MATCHING,
-)
-from wemake_python_styleguide.logic.tokens.constants import NEWLINES
 from wemake_python_styleguide.logic.tokens.newlines import next_meaningful_token
-from wemake_python_styleguide.logic.tokens.queries import only_contains
 from wemake_python_styleguide.logic.tokens.strings import (
     get_docstring_tokens,
     has_triple_string_quotes,
@@ -36,154 +15,12 @@ from wemake_python_styleguide.violations.best_practices import (
     WrongMultilineStringUseViolation,
 )
 from wemake_python_styleguide.violations.consistency import (
-    BracketBlankLineViolation,
-    ExtraIndentationViolation,
     InconsistentComprehensionViolation,
-    WrongBracketPositionViolation,
 )
 from wemake_python_styleguide.visitors.base import BaseTokenVisitor
 from wemake_python_styleguide.visitors.decorators import alias
 
 TokenLines: TypeAlias = DefaultDict[int, List[tokenize.TokenInfo]]
-
-
-@final
-class ExtraIndentationVisitor(BaseTokenVisitor):
-    """
-    Is used to find extra indentation in nodes.
-
-    Algorithm:
-    1. goes through all nodes in a module
-    2. remembers minimal indentation for each line
-    3. compares each two closest lines: indentation should not be >4
-
-    """
-
-    _ignored_tokens: ClassVar[Tuple[int, ...]] = (
-        tokenize.NEWLINE,
-    )
-
-    _ignored_previous_token: ClassVar[Tuple[int, ...]] = (
-        tokenize.NL,
-    )
-
-    def __init__(self, *args, **kwargs) -> None:
-        """Creates empty counter."""
-        super().__init__(*args, **kwargs)
-        self._offsets: Dict[int, tokenize.TokenInfo] = {}
-
-    def visit(self, token: tokenize.TokenInfo) -> None:
-        """Goes through all tokens to find wrong indentation."""
-        self._check_extra_indentation(token)
-
-    def _check_extra_indentation(self, token: tokenize.TokenInfo) -> None:
-        lineno, _offset = token.start
-        if lineno not in self._offsets:
-            self._offsets[lineno] = token
-
-    def _get_token_offset(self, token: tokenize.TokenInfo) -> int:
-        if token.exact_type == tokenize.INDENT:
-            return token.end[1]
-        return token.start[1]
-
-    def _check_individual_line(
-        self,
-        lines: Sequence[int],
-        line: int,
-        index: int,
-    ) -> None:
-        current_token = self._offsets[line]
-        if current_token.exact_type in self._ignored_tokens:
-            return
-
-        previous_token = self._offsets[lines[index - 1]]
-        if previous_token.exact_type in self._ignored_previous_token:
-            return
-
-        offset = self._get_token_offset(current_token)
-        previous_offset = self._get_token_offset(previous_token)
-
-        if offset > previous_offset + 4:
-            self.add_violation(ExtraIndentationViolation(current_token))
-
-    def _post_visit(self) -> None:
-        lines = sorted(self._offsets.keys())
-
-        for index, line in enumerate(lines):
-            if index == 0 or line != lines[index - 1] + 1:
-                continue
-            self._check_individual_line(lines, line, index)
-
-
-@final
-class BracketLocationVisitor(BaseTokenVisitor):
-    """
-    Finds closing brackets location.
-
-    We check that brackets can be on the same line or
-    brackets can be the only tokens on the line.
-
-    We track all kind of brackets: round, square, and curly.
-    """
-
-    def __init__(self, *args, **kwargs) -> None:
-        """Creates line tracking for tokens."""
-        super().__init__(*args, **kwargs)
-        self._lines: TokenLines = defaultdict(list)
-
-    def visit(self, token: tokenize.TokenInfo) -> None:
-        """Goes through all tokens to separate them by line numbers."""
-        self._lines[token.start[0]].append(token)
-
-    def _annotate_brackets(
-        self,
-        tokens: List[tokenize.TokenInfo],
-    ) -> Mapping[int, int]:
-        """Annotates each opening bracket with the nested level index."""
-        brackets = dict.fromkeys(MATCHING, 0)
-        for token in tokens:
-            if token.exact_type in MATCHING.keys():
-                brackets[token.exact_type] += 1
-            if token.exact_type in MATCHING.values():
-                reverse_bracket = get_reverse_bracket(token)
-                if brackets[reverse_bracket] > 0:
-                    brackets[reverse_bracket] -= 1
-        return brackets
-
-    def _check_closing(
-        self,
-        token: tokenize.TokenInfo,
-        index: int,
-        tokens: List[tokenize.TokenInfo],
-    ) -> None:
-        tokens_before = tokens[:index]
-        annotated = self._annotate_brackets(tokens_before)
-        if annotated[get_reverse_bracket(token)] == 0:
-            if not only_contains(tokens_before, ALLOWED_EMPTY_LINE_TOKENS):
-                self.add_violation(WrongBracketPositionViolation(token))
-
-    def _check_individual_line(self, tokens: List[tokenize.TokenInfo]) -> None:
-        for index, token in enumerate(tokens):
-            if token.exact_type in MATCHING.values():
-                self._check_closing(token, index, tokens)
-                if index == 0:
-                    self._check_empty_line_wrap(token, delta=-1)
-            elif token.exact_type in MATCHING and last_bracket(tokens, index):
-                self._check_empty_line_wrap(token, delta=1)
-
-    def _check_empty_line_wrap(
-        self,
-        token: tokenize.TokenInfo,
-        *,
-        delta: int,
-    ) -> None:
-        tokens = self._lines.get(token.start[0] + delta)
-        if tokens is not None and only_contains(tokens, NEWLINES):
-            self.add_violation(BracketBlankLineViolation(token))
-
-    def _post_visit(self) -> None:
-        for tokens in self._lines.values():
-            self._check_individual_line(tokens)
 
 
 @final